requirements.txt
# Requirements without Version Specifiers
numpy
pandas
Pillow
PyOpenGL
PyOpenGL_accelerate
regex
